The existence of this type is due to a long-standing Synapse bug where power levels can be strings (matrix-org/synapse#1237, matrix-org/matrix-doc#3264). However, it gets hit quite hard when resolving room state as we need to pull power levels out of possibly a large set of events.

This PR changes from `json.Unmarshal` to `strconv.ParseInt`/`strconv.ParseFloat` where possible, as they are significantly faster (not relying on `reflect` internally, for a start).

Before this PR:
```
BenchmarkLevelJSONValueInt-12             741554              1430 ns/op
BenchmarkLevelJSONValueFloat-12           317689              3312 ns/op
BenchmarkLevelJSONValueString-12          547220              2123 ns/op
```

With this PR:
```
BenchmarkLevelJSONValueInt-12            1332276               884.9 ns/op
BenchmarkLevelJSONValueFloat-12           585170              1949 ns/op
BenchmarkLevelJSONValueString-12          708974              1757 ns/op
```